+++ packages/model/src/-private/references/has-many.ts
@@ -21,9 +21,8 @@ import type {
 } from '@warp-drive/core-types/spec/raw';
 
 import { assertPolymorphicType } from '../debug/assert-polymorphic-type';
-import { areAllInverseRecordsLoaded, LegacySupport } from '../legacy-relationships-support';
+import { areAllInverseRecordsLoaded, LEGACY_SUPPORT, LegacySupport  } from '../legacy-relationships-support';
 import type ManyArray from '../many-array';
-import { LEGACY_SUPPORT } from '../model';
 
 /**
   @module @ember-data/model
